requirements.txt
requests~=2.28.1
lxml~=4.9.1
python-dotenv~=0.21.0